Refactor to a separate hook
